refactor(frontend): Separate the page routes and sidebar from Navbar
Navbar is too serving too big of a purpose, it should remain a Navbar component only.

- Moved the sidebar to its own component
- Moved the routes to the app component
