+++ frontend/src/components/layout/Navbar.jsx
@@ -1,95 +1,40 @@
-import React, { useState } from "react";
-import {
-  Box,
-  Button,
-  Container,
-  Flex,
-  HStack,
-} from "@chakra-ui/react";
-import { motion } from "framer-motion";
+import { Box, Button, Flex, HStack, Input } from "@chakra-ui/react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBars } from "@fortawesome/free-solid-svg-icons";
+import { faBars, faSearch } from "@fortawesome/free-solid-svg-icons";
 import { faUserCircle } from "@fortawesome/free-regular-svg-icons";
-import { Route, Routes } from "react-router-dom";
-import Dashboard from "../../pages/Dashboard";
-import CalendarPage from "../../pages/CalendarPage";
-import ManageList from "../../pages/ManageList";
-import Settings from "../../pages/Settings";
-import SButtons from "./SidebarButtons";
+import { InputGroup } from "../snippets/input-group";
 
-const MotionBox = motion(Box);
-
-const Navbar = () => {
-  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
-
-  const toggleSidebar = () => {
-    setIsSidebarOpen(!isSidebarOpen);
-  };
-
+const Navbar = ({ toggleSidebar }) => {
   return (
-    <Flex flexDirection="column" h="100vh" bg="black">
-      <Container maxW="full" px={4} bgColor="gray.200">
-        <Flex
-          h="50px"
-          alignItems="center"
-          justifyContent="space-between"
-          flexDir={{
-            base: "column",
-            sm: "row",
-          }}
-        >
-          <HStack gap={3}>
-            <Button variant="plain" onClick={toggleSidebar}>
-              <FontAwesomeIcon icon={faBars} size="2xl" color="red" />
-            </Button>
-          </HStack>
-          <FontAwesomeIcon icon={faUserCircle} size="2xl" color="black" />
-        </Flex>
-      </Container>
-      <Flex flex={1} h="calc(100vh - 50px)" overflow="hidden" position="relative">
-        <MotionBox
-          w="250px"
-          bg="blue.900"
-          color="white"
-          p={4}
-          boxShadow="md"
-          h="100%"
-          position="absolute"
-          zIndex="overlay"
-          left="0"
-          top="0"
-          initial={{ x: "-100%" }}
-          animate={{ x: isSidebarOpen ? "0%" : "-100%" }}
-          exit={{ x: "-100%" }}
-          transition={{
-            type: "spring",
-            stiffness: 100,
-            damping: 18,
-          }}
-        >
-          <SButtons />
-        </MotionBox>
-        <Box
-          flex={1}
-          p={6}
-          ml={isSidebarOpen ? "250px" : "0"}
-          transition="margin-left 0.5s ease"
-          bg="gray.100"
-          h="100%"
-          w="100%"
-          position="relative"
-          zIndex="base"
-        >
-          <Routes>
-            <Route path="/Dashboard" element={<Dashboard />} />
-            <Route path="/CalendarPage" element={<CalendarPage />} />
-            <Route path="/ManageList" element={<ManageList />} />
-            <Route path="/Settings" element={<Settings />} />
-          </Routes>
-        </Box>
+    <Box w="full" px={4} bgColor="gray.300">
+      <Flex
+        h="50px"
+        alignItems="center"
+        justifyContent="space-between"
+        flexDir={{
+          base: "column",
+          sm: "row",
+        }}
+      >
+        <HStack gap={3}>
+          <Button variant="plain" onClick={toggleSidebar}>
+            <FontAwesomeIcon icon={faBars} size="2xl" color="red" />
+          </Button>
+					<Box>
+            <InputGroup
+              flex="1"
+              endElement={
+                <FontAwesomeIcon icon={faSearch} size={"md"} color={"black"} />
+              }
+            >
+              <Input placeholder="Search..." variant={"outline"} size={"sm"}/>
+            </InputGroup>
+          </Box>
+        </HStack>
+        <FontAwesomeIcon icon={faUserCircle} size="2xl" color="black" />
       </Flex>
-    </Flex>
+    </Box>
   );
 };
 
-export default Navbar;  
+export default Navbar;
